+++ .travis.yml
@@ -1,8 +1,11 @@
 language: go
 
 go:
-  - 1.6
+  - 1.8
 
 script:
-  - cd provider-dnspod
-  - go build
+  - sudo add-apt-repository ppa:masterminds/glide -y && sudo apt-get update
+  - sudo apt-get install glide -y
+  - glide up
+  - make build
+  - make test
