+++ scripts/errcheck.sh
@@ -1,0 +1,25 @@
+#!/usr/bin/env bash
+
+# Check gofmt
+echo "==> Checking dnspod provider for unchecked errors..."
+echo "==> NOTE: at this time we only look for uncheck errors in the dnspod package"
+
+if ! which errcheck > /dev/null; then
+    echo "==> Installing errcheck..."
+    go get -u github.com/kisielk/errcheck
+fi
+
+err_files=$(errcheck -ignoretests -ignore \
+  'github.com/hashicorp/terraform/helper/schema:Set' \
+  -ignore 'bytes:.*' \
+  -ignore 'io:Close|Write' \
+  ./dnspod/...)
+
+if [[ -n ${err_files} ]]; then
+    echo 'Unchecked errors found in the following places:'
+    echo "${err_files}"
+    echo "Please handle returned errors. You can check directly with \`make errcheck\`"
+    exit 1
+fi
+
+exit 0
